[WIP] sh-bisect demo for searching branch deletion revision
Given the a path of a delted SVN branch like "$REPO/branches/experiment@42"
the usual svn commands can only operate on revisions older than 42. Also
the svn commands do not show in whith revision the branch has been deleted.
Greping `svn log -v` is also not always possible, because the branch could
have been deleted by deleting some parent branch.

The only obvious way is to do a binary search between the last known revision
and HEAD. svn will error out on command like this when the branche in the upper
limit is not same as the given branch.

	svn log --stop-on-copy --limit 1 -r80:42 $REPO/branches/experiment@42

This could be the bisection predicate.